#10 @othman853: Restructuring the application folders
