+++ src/database.js
@@ -1,10 +1,10 @@
 const mongoose = require('mongoose')
 
-module.exports = {
-  connect: () => mongoose.connect(process.env.MONGODB_URI),
+module.exports = (databaseUri) => ({
+  connect: () => mongoose.connect(databaseUri),
 
   disconnect: () => new Promise((resolve, reject) =>
     mongoose.connection.close((error) => error
       ? reject(error)
       : resolve()))
-}
+})
